refactor: split tactics into individual files per tactic
Split tactics into individual files organized by time of day (day/night)
to prevent merge conflicts and enable parallel development.

Structure:
- packages/shared/src/tacticIds/{day,night}/ - Individual tactic ID files
- packages/core/src/data/tactics/{day,night}/ - Individual tactic card definitions
