+++ packages/shared/src/tacticIds/day/earlyBird.ts
@@ -1,0 +1,1 @@
+export const TACTIC_EARLY_BIRD = "early_bird" as const;
