+++ packages/shared/src/tacticIds/day/greatStart.ts
@@ -1,0 +1,1 @@
+export const TACTIC_GREAT_START = "great_start" as const;
