+++ packages/shared/src/tacticIds/day/index.ts
@@ -1,0 +1,26 @@
+/**
+ * Day tactic IDs
+ */
+
+export { TACTIC_EARLY_BIRD } from "./earlyBird.js";
+export { TACTIC_RETHINK } from "./rethink.js";
+export { TACTIC_MANA_STEAL } from "./manaSteal.js";
+export { TACTIC_PLANNING } from "./planning.js";
+export { TACTIC_GREAT_START } from "./greatStart.js";
+export { TACTIC_THE_RIGHT_MOMENT } from "./theRightMoment.js";
+
+import { TACTIC_EARLY_BIRD } from "./earlyBird.js";
+import { TACTIC_RETHINK } from "./rethink.js";
+import { TACTIC_MANA_STEAL } from "./manaSteal.js";
+import { TACTIC_PLANNING } from "./planning.js";
+import { TACTIC_GREAT_START } from "./greatStart.js";
+import { TACTIC_THE_RIGHT_MOMENT } from "./theRightMoment.js";
+
+export const DAY_TACTIC_IDS = [
+  TACTIC_EARLY_BIRD,
+  TACTIC_RETHINK,
+  TACTIC_MANA_STEAL,
+  TACTIC_PLANNING,
+  TACTIC_GREAT_START,
+  TACTIC_THE_RIGHT_MOMENT,
+] as const;
