+++ packages/shared/src/tacticIds/day/manaSteal.ts
@@ -1,0 +1,1 @@
+export const TACTIC_MANA_STEAL = "mana_steal" as const;
