+++ packages/shared/src/tacticIds/day/planning.ts
@@ -1,0 +1,1 @@
+export const TACTIC_PLANNING = "planning" as const;
