+++ packages/shared/src/tacticIds/day/rethink.ts
@@ -1,0 +1,1 @@
+export const TACTIC_RETHINK = "rethink" as const;
