+++ packages/shared/src/tacticIds/day/theRightMoment.ts
@@ -1,0 +1,1 @@
+export const TACTIC_THE_RIGHT_MOMENT = "the_right_moment" as const;
