+++ packages/shared/src/tacticIds/night/fromTheDusk.ts
@@ -1,0 +1,1 @@
+export const TACTIC_FROM_THE_DUSK = "from_the_dusk" as const;
