+++ packages/shared/src/tacticIds/night/index.ts
@@ -1,0 +1,26 @@
+/**
+ * Night tactic IDs
+ */
+
+export { TACTIC_FROM_THE_DUSK } from "./fromTheDusk.js";
+export { TACTIC_LONG_NIGHT } from "./longNight.js";
+export { TACTIC_MANA_SEARCH } from "./manaSearch.js";
+export { TACTIC_MIDNIGHT_MEDITATION } from "./midnightMeditation.js";
+export { TACTIC_PREPARATION } from "./preparation.js";
+export { TACTIC_SPARING_POWER } from "./sparingPower.js";
+
+import { TACTIC_FROM_THE_DUSK } from "./fromTheDusk.js";
+import { TACTIC_LONG_NIGHT } from "./longNight.js";
+import { TACTIC_MANA_SEARCH } from "./manaSearch.js";
+import { TACTIC_MIDNIGHT_MEDITATION } from "./midnightMeditation.js";
+import { TACTIC_PREPARATION } from "./preparation.js";
+import { TACTIC_SPARING_POWER } from "./sparingPower.js";
+
+export const NIGHT_TACTIC_IDS = [
+  TACTIC_FROM_THE_DUSK,
+  TACTIC_LONG_NIGHT,
+  TACTIC_MANA_SEARCH,
+  TACTIC_MIDNIGHT_MEDITATION,
+  TACTIC_PREPARATION,
+  TACTIC_SPARING_POWER,
+] as const;
