+++ packages/shared/src/tacticIds/night/longNight.ts
@@ -1,0 +1,1 @@
+export const TACTIC_LONG_NIGHT = "long_night" as const;
