+++ packages/shared/src/tacticIds/night/manaSearch.ts
@@ -1,0 +1,1 @@
+export const TACTIC_MANA_SEARCH = "mana_search" as const;
