+++ packages/shared/src/tacticIds/night/midnightMeditation.ts
@@ -1,0 +1,1 @@
+export const TACTIC_MIDNIGHT_MEDITATION = "midnight_meditation" as const;
