+++ packages/shared/src/tacticIds/night/preparation.ts
@@ -1,0 +1,1 @@
+export const TACTIC_PREPARATION = "preparation" as const;
