+++ packages/shared/src/tacticIds/night/sparingPower.ts
@@ -1,0 +1,1 @@
+export const TACTIC_SPARING_POWER = "sparing_power" as const;
